fix: move parsing logic to store, setup for hadnleing multipart
